+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix mkCljCli helper function
+- Add support for Leiningen projects
 
 ## 0.2.0 (2022-06-13)