+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix mkCljCli helper function
 - Add support for Leiningen projects
 - Add lockfile option to mkCljBin
+- Add mkBabashka
 
 ## 0.2.0 (2022-06-13)