+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add support for Leiningen projects
 - Add lockfile option to mkCljBin
 - Add mkBabashka
+- Add `bbTasksFromFile`
 
 ## 0.2.0 (2022-06-13)